Moving eloquent tests into Unit Tests
