+++ phpunit.xml
@@ -16,10 +16,6 @@
         <testsuite name="Unit">
             <directory suffix="Test.php">./tests/Unit</directory>
         </testsuite>
-
-        <testsuite name="Eloquent">
-            <directory suffix="Test.php">./tests/Eloquent</directory>
-        </testsuite>
     </testsuites>
     <filter>
         <whitelist processUncoveredFilesFromWhitelist="true">
