+++ tests/Unit/Eloquent/AdminTest.php
@@ -1,38 +1,36 @@
 <?php
 
-namespace Tests\Eloquent;
+namespace Tests\Unit\Eloquent;
 
 use App\Authenticatable\Admin;
-use App\Authenticatable\Assistant;
 use App\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
-class AssistantTest extends TestCase
+class AdminTest extends TestCase
 {
     use RefreshDatabase;
 
     /**
-     * @var Assistant
+     * @var Admin
      */
-    protected $assistant;
+    protected $admin;
 
     protected function setUp()
     {
         parent::setUp();
 
-        $this->assistant = create(Assistant::class);
+        $this->admin = create(Admin::class);
     }
 
     /**
      * @test
-     * @group f
      */
-    public function assistant_does_not_include_users()
+    public function admin_does_not_include_users()
     {
         create(User::class);
 
-        $this->assertEquals(1, Assistant::count());
+        $this->assertEquals(1, Admin::count());
         $this->assertEquals(2, User::count());
     }
 }
