+++ tests/Unit/Eloquent/AssistantTest.php
@@ -1,36 +1,37 @@
 <?php
 
-namespace Tests\Eloquent;
+namespace Tests\Unit\Eloquent;
 
 use App\Authenticatable\Admin;
+use App\Authenticatable\Assistant;
 use App\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
-class AdminTest extends TestCase
+class AssistantTest extends TestCase
 {
     use RefreshDatabase;
 
     /**
-     * @var Admin
+     * @var Assistant
      */
-    protected $admin;
+    protected $assistant;
 
     protected function setUp()
     {
         parent::setUp();
 
-        $this->admin = create(Admin::class);
+        $this->assistant = create(Assistant::class);
     }
 
     /**
      * @test
      */
-    public function admin_does_not_include_users()
+    public function assistant_does_not_include_users()
     {
         create(User::class);
 
-        $this->assertEquals(1, Admin::count());
+        $this->assertEquals(1, Assistant::count());
         $this->assertEquals(2, User::count());
     }
 }
